Map IdP assertions to local; initial code version
A federated Identity Provider (IdP) provides assertions (i.e. claims)
regarding the subject (i.e. principal) after a successful
authentication. Because a federated IdP is by definition external to
ODL the assertions must be mapped from the IdP into local values.

The mapping is performed by evaluating a series of rules expressed in
JSON notation.

This package implements a RuleProcessor class which accepts a set of
rules and an assertion. It emits the transformed assertions as JSON
object. JSON is the exchange format for the rules, assertion, and
result. As such the package includes the IdpJson class which
transforms between JSON and the Java data structures used by the
RuleProcessor.

There is complete documentation for using the RuleProcessor which will
be added in a later commit.

Change-Id: I707c2b7dd5be381ef25f2fcdfb1c73481a63c9e5
Signed-off-by: John Dennis <jdennis@redhat.com>