Remove some captured values, add comment, add logging
CGI_CONTENT_TYPE and CGI_CONTENT_LENGTH are not relevant to an
authentication claim, so remove them.

Reorder the CGI variable name declarations so they are sorted. Makes
it easier to cross check their usage when things are in the same
order.

Remove the wholesale collection of HTTP headers since in general
they cannot be trusted, but retain the selected capture of specific
HTTP headers configured by an admin.

Add comment explaining why getAttributeNames() has problems and how we
adopt.

Add logger and then log the claims map.

Change-Id: I19a274601eceb5e2f24a3c055e9c73d4bb52e9b9
Signed-off-by: John Dennis <jdennis@redhat.com>