Update patch set 1
Patch Set 1:

Hi John,

While reviewing this, I wonder if we can just make sure that the ClaimAuthFilter only trusts the loopback address?  Or do we expect the Apache server to be on a separate machine than the controller, in which case then we cannot limit the jetty connector for federation to just localhost?

Patch-set: 1
Label: Code-Review=0