Update patch set 2
Patch Set 2:

We expect the common deployment to be Apache and the app server to be running on the same host, therefore restricting proxy auth to the loopback is an option that at first blush would appear to make sense. However that is only one possible deployment option.

It's not unusual in large enterprises to proxy authentication to other servers. Or for that matter you may want to load balance. Or you may want to (or be forced by auditing requirements) to use SSL (which doesn't work on loopback). Any of the above would require accepting requests from an address other than the loopback. So my thinking is it's best to give admin in charge of deployment the option to configure the relationship between the two servers as best fits the needs of his organization and to establish the needed trust via port separation.

I'm open to counter arguments, it can be easy to fail to consider all the issues.

Patch-set: 2
Reviewer: Gerrit User 2084 <2084@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0