Bug 1964
Several fixes were made in this commit:

1) Separate the federation endpoint into its own webapp (aaa-authn-federation), so it can bind to a separate authorized proxy port.
2) Move initialization code in SssdClaimAuth from the constructor to the init() method to make sure it is initialized via OSGi lifecycle (fixed OSGi loading issue of SssdClaimAuth)
3) Clean up superflous log.info() messages from IdmLight
4) Fix ClaimAuthFilter to emit 401 error right away if we are federating on a non-authorized proxy port.
5) Add basic integration tests (-Paaa-it) for IdMLight and federation.
6) Configure:
   a) IdmLight APIs (/auth/*) to listen on "adminConn" Jetty connector.
   b) Federation API (/oauth2/federation/) to listen on "federationConn" Jetty connector.

Note:  Currently, the aforementioned Jetty connectors are NOT configured on the ODL controller, so that means those APIs in 6) are not available by default.
To activate them, the sample jetty.xml under aaa-it/src/test/resources should be copied over to the controller's assembly/etc/jetty.xml.  The sample jetty.xml
enables the adminConn on port 8282, localhost only, and the federationConn on port 8383.  So, for example, a POST to the federation endpoint would be:

curl -i -XPOST http://localhost:8383/oauth2/federation/

Change-Id: I1bc939536806d864e462b5cd0f69d1bb1777058d
Signed-off-by: Liem Nguyen <liem_m_nguyen@hp.com>