Update patch set 1
Patch Set 1:

Looks good, John... But, did you forget to include it in the web.xml of the aaa-authn-federation project?  And the filter should probably lives there too.  Since I think this is core functionality to AAA for working with non-AJP protocol, can we change the X-SSSD to X-AAA?  :)

Thanks,
Liem

Patch-set: 1
Label: Code-Review=0