Update patch set 1
Patch Set 1:

Not including it in web.xml was deliberate. I'm not sure if that was right or wrong. My thinking was it was unnecessary if you're using AJP, although it should be benign (it would just chew up unnecessary cycles). But since it looks like we're committed to using Jetty it should be there (perhaps at least present but inside a comment). Also if you're not doing Fedration with SSSD it is also just a waste of cycles.

I think the right thing to do is add it but put it inside a comment that can easily be uncommented later.

X-SSSD vs. X-AAA, my thinking is theses headers should be identified as belonging to SSSD because that's where they originate from. If they were labeled AAA that doesn't really help someone understand what these mysterious headers are other than AAA will consume them, I think it's better to label things using the single producers name rather than the consumer (which in theory could be multiple consumers). Plus I have an ulterior motive, I want to use the same doc on how to configure Apache with SSSD for projects other than ODL :-)

I'll add the web.xml filter and resubmit.

Patch-set: 1
Label: Verified=0