Do not instantiate Booleans, Strings and Longs
The two possible values are available as constants, so use them
directly, lowering the amount of garbage we generate.

For an empty String, we can just use a literal. For Longs, instead of
explicit construction, we can use valueOf(), which can end up using the
same instance. Also, we can use autoboxing to let JVM promote a value as
appropriate.

Change-Id: Icf9405d691a08c20dde78a5fe05bd4fab5947741
Signed-off-by: Robert Varga <rovarga@cisco.com>