Ensure that the TokenStore is unregistered when closed
The ServiceLocator maintains a reference to all registered token stores and
when a request for get is made it goes through all the registered token stores
to see which one might have the token.

If we have a stale token store then it will also have a reference to a stale
data broker which will throw an exception when a transaction is created on it -
this exception happens to be a runtime exception and if ignored will result in
the exception surfacing all the way to the filter.

Change-Id: I4a66b144529ab5a99a6f53e5bb2ab176e18de703
Signed-off-by: Moiz Raja <moraja@cisco.com>