Update patch set 8
Patch Set 8: Code-Review+1

Looks good as a fix. It is now in a shape where you can use odbc and grown up databases properly.

We can change it to be configurable later on for Be. This will do for now.

Patch-set: 8
Reviewer: Gerrit User 2774 <2774@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1