Update patch set 3
Patch Set 3:

...:o)
Silly me...I wrote DataEncrypter...
I didn't realize that PBKDF2WithHmacSHA1 is an encryption, I though it was a stronger hashing than MD5.
As a thumb rule, one should never use password encryption but password hash, e.g. one way encryption so you can never decrypt the password. In the case when the user forgets his password, you create a procedure to reset the password and that way the admin can never know the user password and can only know if it was previously used by keeping the history of the hashes.

Patch-set: 3