Bug 4022 AAA does not enforce domain-unique user names
Adds a check to grant creation to ensure another user with the same name isn't
already granted permission to the domain.  Thus, if there is a domain called
"sdn", and a user called "admin" (userid=1) is granted permissions to
administer sdn, then afterwards a user also called "admin" (userid=3) is not
allowed to obtain a grant to administer "sdn".

Change-Id: Id338018b82d9415e36115384b024ce2fb1365003
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>