Bug 4022 AAA does not enforce unique user names
Other parts of AAA code relies upon unique username.  An example of such code
is IdmLightProxy.dbAuthenticate() which queries for a username but utilizes
only the first returned User using "user=userList.get(0);".  This change
enforces a unique user name in SQL, and thus attempts to insert duplicate
users with the same name will result in an SQLException.  This approach
was taken since current AAA code does not utilize transactions, and
a "SELECT" followed by an "INSERT" is not atomic.  Thus, it was decided to
enforce uniqueness at the database layer.

Change-Id: Id338018b82d9415e36115384b024ce2fb1365003
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>