Update patch set 1
Patch Set 1:

> I would prefer if we enforced uniqueness of username+domain-name
 > (ie joe@coke.com and joe@pepsi.com should be allowed). That beings
 > said, we do need to have a non domain affiliated user that's also
 > unique, eg admin. Maybe a way to do that is to have a "@local"
 > domain.

Right now the admin user is scoped to the "sdn" domain.  I don't know whether we want to keep it this way going forward?

I have created code to enforce the relationship you described in your last comment.  I have to warn that this relationship is significantly more complicated, and thus the code used to enforce the relationship is more involved.  I do agree that multi-tenancy is a real use case we need to consider (along with ryan@coke and ryan@sprite).

Patch-set: 1