Update patch set 2
Patch Set 2:

Please let me know what you think of this patch.  I have tested it manually and it seems to work well for me.  Perhaps this is a case where we can include an automated test in the test plan Carmen is working on.  I'll ad him to the review to see if he has any ideas!

Patch-set: 2