Bug 4042 AuthN fails for users that aren't granted permission with default sdn domain
Prior to these changes, basic authentication assumes that the provided
credentials are attempting AuthN against the default domain.  This is not
always the case, since a user does not need to be part of the "sdn" domain,
and still needs a method to authenticate and access protected resources.  This
change adds the capability to pass a queryParameter to a REST URL in the
following way:

curl -u cokeadmin:pass "http://localhost:8181/auth/v1/users?domainname=coke"

If the parameter is not present, then the default domain is assumed.  This is
the traditional case:

curl -u admin:admin http://localhost:8181/auth/v1/users

The URL will accept any number and variety of query paramters, but only the
first occurrence of "domainname" will affect the end behavior.  If a
non-existent domainname is used as input, then a 401 is returned since the
user does not have a grant for the non-existent domain.

Change-Id: Idd87d944ef5b841c01fc67e96f11179936b48fb4
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>