Update patch set 4
Patch Set 4: -Code-Review

I believe this documentation is ready for a review.  The week before the ODL Summit we discussed moving this into the README.md file.  However, upon reading that file a little more closely, this seems too low-level and instructional for that file.  I believe that we should include this as-is in a separate location since it is much more in-depth and hands-on.

Patch-set: 4
Label: Code-Review=0