Update patch set 1
Patch Set 1:

> Any thoughts on calling on AuthZServiceImpl directly (which the
 > sample implementation did) vs using the RPC interface to invoke an
 > AuthZ decision?

Is this in order to standardize on an RPC interface;  I remember you bringing this idea up in our weekly meeting?  I don't know what it buys us in this case...  but I can look into it if you think it is a good idea.

The part I am a little more concerned about is whether we have captured the correct "action-type" enum in the authorization-schema.yang.  I think we should probably use a separate patch to convert from enum since enums are not extensible or future-proof.

Patch-set: 1
Label: Verified=0