Update patch set 8
Patch Set 8:

Tony, any idea why it might be failing? 

As far as I can tell I changed all instances of netconf to the new one but the builder refuses to pick up the netconf groupid. If I purposely change the version in authn/pom.xml to 0.4.0 then it picks up the right groupid but fails for obvious reasons.

Patch-set: 8