Update patch set 4
Patch Set 4:

Seems to be running into the same issues as my attempt to do this task:

    https://git.opendaylight.org/gerrit/26110

I'm not sure why it's not processing the features.xml correctly and getting the old one. My hunch is that it's an order of operations issue. Where Karaf is pulling the old SNAPSHOT from nexus.opendaylight.org while the features/authn version that is built hasn't completed building yet so has yet to deploy the artifact or install it in the local repo.

Patch-set: 4