Update patch set 4
Patch Set 4:

To prove this point. You can cause the tests to pass if you first run:

 mvn clean install -DskipTests
 mvn clean install

By skipping tests on the first run Maven will install the new feature to your local repo then when you run Maven the 2nd time with tests enabled it passes because it is now using the newest version of the package.

Patch-set: 4