Update patch set 1
Patch Set 1:

We had a similar case to this with authn earlier this year; this one may have been missed.  My only concern is the features.xml seems to have been moved to the features folder from the resources folder.  Seeing if simply moving the file back fixes this without having to put the qualified (non-variable) path.

Patch-set: 1