Update patch set 13
Patch Set 13:

(28 comments)

In a high level, i am not done yet. Still need to extract the H2 to a different project and create a service locator so the store will be dynamically selected.

The method compareObject is used to encapsulate the null object checks and when both are not null it is delegating the check to the object equal method.

Patch-set: 13
Label: Verified=0