Update patch set 15
Patch Set 15:

I am using reflection here to avoid a lot of code duplication and make it easy to implement new model objects and change model objects. I am well aware of reflection advantages and disadvantages and in this case the reflection is being used privately to the class, hence there is no issues with code trace-ability. As for performance, this is a persistence service and as such the reflection performance is by factor neglectable vs. any persistence operation.

Patch-set: 15
Label: Verified=0