DO NOT MERGE: WIP: Bug5760 AAA Jersey and JAX-RS upgrades
Upgrade versions of Jersey and JAX-RS.  This is a first pass, and since the testing
framework code has not yet been updated, the tests have been temporarily removed.
These tests will be brought back in again before this patch is merged.

This patch is here to save work off at a checkpoint.

Change-Id: Ic12bfe81feb18ad00221f39e3d973978b57bb7ec
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>