Update patch set 22
Patch Set 22:

We should probably add an entry to the authz config to lock down access to this model.  Take a look at aaa-authz-config/src/main/resources/initial/08-authz-config.xml;  Wojciech you initially authored this... is there a way to include another default rule?  Still reading over the rest of this; just a general comment to prep this for production.

Patch-set: 22
Label: Verified=0