Update patch set 3
Patch Set 3:

> (1 comment)

The enum pattern is fine but a bit awkward in my opinion.

The book was quoted not for the enum pattern, but for the public instance variables that are accessed all over the place.  This violates the principal of data encapsulation, a major advantage IMHO for OO programming.

And there certainly are multiple threads acting on some of these.  When the feature is installed, then the *Auth instance is added to the collection.  At the same time, it could be referenced from a REST call trying to authenticate a token/claim.  Or another scenario, where 2 features are installed in close time span with one another and are both trying to manipulate the TokenAuth collection.

Patch-set: 3
Reviewer: Gerrit User 3635 <3635@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0