Update patch set 3
Patch Set 3: Code-Review+1

In terms of getters and setters, there's a long history of discussion if not several books written (about them being an anti pattern). For many, incl me, they're a useful tool in general but in this particular case where the object's sole intent is to allow service injection of a specific type, and hold it, not a vital need.

Re threads; yes correct multiple access is indeed possible, however enum objects are inherently thread safe, while for the list item which was the only risky one there, you added the Vector class which made it thread safe (and it in theory doesn't need a synchronised getter/setter).

In summary; no major objections to the changes, but I'd say that, Vector aside, the rationale falls more on style.

Patch-set: 3
Reviewer: Gerrit User 2106 <2106@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1