Update patch set 2
Patch Set 2:

(3 comments)

While I have no problems with this great testing on the IDP portion of the code, I do want to give you the heads up as it has been discussed many times in our weekly meetings that we plan to deprecate a lot of this code path in favor of using Apache Shiro for IdP connections.  If  you attend our weekly meetings, you will have a bit better of an idea about the roadmap.  Please fix the formatting issues and I will merge this!

Patch-set: 2
Label: Code-Review=0