Adds Shiro Web-based authorization capabilities
Shiro supports RBAC for url patterns that may contain asterisks.  Some DOM
leaves should be filtered for security reasons, such as those containing AAA
data.  This allows installation of a set of default, immutable filters;  the
alternative was to allow installation in shiro.ini, but that file is mutable.

Since aaa-shiro isn't presently active, no changes will occur for external
projects.  This is setting up for when the aaa implementation is swapped out
for an Apache Shiro based one.

Change-Id: If0d037cafb90d8cf51713e538f41967f85f325ae
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>