Update patch set 1
Patch Set 1:

AuthZ was working in Lithium... it looks like something may have broken it.  The other plan was to use Shiro, but that is currently delayed since Jersey/JAX-RS changes needed to be reverted.  We can discuss how to move forward on the call;  these changes are good IMO.  Just need to figure out how we should protect the endpoint;  there may be some happy medium.  Thanks for your hard work!

Patch-set: 1