Update patch set 6
Patch Set 6:

This patch needs a much better message explaining what it does, how and why. Overall this looks like a discovery service. We already have gossip in clustering, why do we have a custom one?

The API and implementation are not separated as per OSGi best practices.

Can we discuss it tomorrow or on next week's clustering call?

Patch-set: 6