Update patch set 6
Patch Set 6:

(16 comments)

Hi Robert,
Thanks for review this patch set, I have applied some of your comments. However this is half backed and I am not finished yet. I am missing a journal and a lot of unit for completion.

This activity is coming to overcome bug:
Bug 4702 - Sometimes getting "SchemaResolutionException" exception when starting controller.

Which is blocking us from reaching out commit of AAA clustering.

It is built to give some clustering ability to the H2 store implementation to satisfy the clustering commit we have for AAA clustering until the above bug is resolved.

Patch-set: 6
Label: Verified=0