Update patch set 6
Patch Set 6:

(2 comments)

I would opt for using off-the-shelf solutions rather than implementing your own clustered datastore - it's a lot of work, especially to try to cram in after the freeze. Folks spend many months/years on clustered frameworks to make them robust (case in point - the mdsal datastore). 

Another option is to fix the issue in Bug 4702 so usage of the md-sal datastore can be re-enabled.

Patch-set: 6