Update patch set 5
Patch Set 5:

> It should at least use a ServiceTracker. But let's get this moved
 > ahead - I can push another patch for that.

Yes;  agreed Tom.  Time was of the essence on this one since w/o proper StoreBuilder initialization, users may experience 401 despite providing valid credentials.  This patch is a stopgap;  it ensures that we wait for the backing IdM store on a separate thread and then calls StoreBuilder.init() when it is found.  I appreciate your help on this one!

Patch-set: 5