Update patch set 7
Patch Set 7: Code-Review-1

This isn't just a light-weight token synchronization method like we discussed last Thursday's AAA weekly meeting.  This is scaffolding for clustered store, which we already have (MD-SAL).  

There are two principal drawbacks of using our current h2 idmlight implementation:
1) Synchronization of credentials across the cluster must be handled manually.  I am fine with this since it is not a regression from last release.  It works OOB and is sufficient.
2) OAuth2 tokens are not clustered across the store.  This is also not a regression.  I am fine with mandating Basic Authentication for clustered AAA in Beryllium;  it is a lot safer than introducing a multi-thousand line patch.

You delivered an optional MD-SAL store;  it is very unfortunate that we are running into config subsystem issues, but I still see this as a win, and I believe we met our requirement per the release plan.  On most powerful systems, the symptom isn't noticed;  and simply put it seems like the CSS bug isn't going to be an easy fix (will need to be put off for future release).  

If you disagree that we did not meet our requirement of delivering an mdsal store, albeit optional, I believe it is safer at this point to alert the team that MD-SAL requirement was not met, rather than trying to push in a bunch of code after freeze.  I personally do not have the cycles to test all of this.

Patch-set: 7