Update patch set 7
Patch Set 7:

> Based on my experience with the config system, I would say it would
 > definitely not be easier. Advertising and consuming services via
 > the config system is not trivial.  Blueprint would be easiest....

This works well;  there may be some changes in how ODL handles OSGi configuration in the future so we should hold off with converting the rest of the bundles for now.  This is a real issue though;  I got a message that someone ran into this the other day.  This patch is a good, quick fix, and since we have employed this tactic before, we know it will work.

Patch-set: 7