Update patch set 10
Patch Set 10:

In looking at the code some more, the 3 Store instances are singletons and created and owned by the H2Store. Therefore I suggest to add a specific close method to each and get rid of the finalizer. Also the classes and methods can all be package-scoped so they're not leaked outside of H2Store.

They are accessed via a rest endpoint so they need to be synchronized wrt to the dbConnection. I think abstracting dbConnection to a base class will make that easier.

Also H2Store has static CRUD methods that create local Store instances. These are only used by unit tests. These should be moved to the unit tests.

Patch-set: 10