Formatting applied to DomainHandler
This change tackles some vital formatting issues with DomainHandler.  The
following is done to the DomainHandler class to promote clarity and conformance
to project style:

1) Move the class documentation comment to the right location.  Add some
detailed information to the class comment concerning how to use the REST
endpoints for DomainHandler.

2) Add function documentation comments explaining their responsibilities.

3) Format in eclipse by clicking "format".

4) Fix any errors reported by "mvn checkstyle:checkstyle".

No actual code or functionality changes were made.  This is strictly a
formatting related patch.

Change-Id: Ie2d0ecff913afbceb280950e1a8d40a43a3958a0
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>