Update patch set 16
Patch Set 16:

The work still in progress, and the idea is at the distribution compilation time, the disto will generate its own keystore and certificate based on pre-define configuration by the user. and at runtime we might us the java keystore API to let the user have this functionality.

Patch-set: 16
Label: Verified=0