Update patch set 1
Patch Set 1:

Additionally, it would be nice to have a Unit Test for this.  This can likely be done using just JUnit, but may be complicated since this is a Jersey Application.  If we cannot use pure JUnit, we can punt on it and add it to the integration test suite.

Patch-set: 1