Bug 5474 Accounting Log for Un/Successful Auth Attempts
Accounting is currently limited to karaf log output messages, which can be
copied to an external syslog server.  Hitherto, AAA plugin didn't report
failed v.s. successful authentication attempts.  This change provides the
ability to enable audit events for successful and unsuccessful authentication
attempts.  This behavior is disabled by default in order to prevent flooding
karaf logs, but may be enabled if an operator feels this logging is
important.

To enable Un/Successful logging, from the karaf shell just type:
> log:set DEBUG org.opendaylight.aaa.shiro.filters.AuthenticationListener

A good deal of testing was added in order to ensure that audit events are
reported in a sane manner.  A utilities class was developed which is
responsible for forming audit log messages.  A custom slf4j appender was added
for testing only in order to track karaf log output, and ensure that audit
events are properly logged.

Change-Id: I21b8dc4ef5b137cf7f968c284a6725da7b02134a
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>