Add support for generic JDBC for AAA
Just wraps the default Shiro implementation.  This allows for enhanced logging
as well as aggregation of all realm implementations to a single package,
making it easier for importing Servlets.

Instructions on how to set up and configure the realm are included right in the
shiro initialization file, shiro.ini.

This abstraction is particularly useful for systems integrators who wish to
leverage an existing JDBC-supporting system for ODL AAA.

Change-Id: I58257a4704e9c302689ef46155972c5ce06dd155
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>