Remove the odl-aaa-keystone-plugin feature
Since this feature doesn't do anything, the AAA team has chosen to directly
remove it.  Since the feature never did anything, there is no need to wait
the extra release cycle.  The advantage to removing this earlier is less
queries surrounding a feature which doesn't work.  Prior to this commit,
the inclusion of this non-functional feature was misleading.

Change-Id: I24136b81dda6a45b13e6edccfb9ffac4468f83bb
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>