Update patch set 1
Patch Set 1: Code-Review-2

Cross bundle dependencies are added to the parent pom if they are local to a project and don't need to be considered across different projects... after all there are some dependencies used in AAA that are not needed across projects.  Local parents are also a good stepping stone in the case that a change is needed in the root parent (odlparent) since you can start off by just making change to the local project and not worry about getting patches merged in any particular order.

Having completed a number of dependency upgrades in AAA and odl as a whole, I do not think this is necessary for the AAA project.

Patch-set: 1
Reviewer: Gerrit User 3635 <3635@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2