Update patch set 1
Patch Set 1:

> aaa-parent is really useful for projects as aaa has different implementation for the same function

Can you elaborate because I really don't get that. If you have different implementation for the same function, then this should be tied to a bundle, thus not to a parent.

> the last test dependancy conflict was a good example for that
Again, I'm not quite sure to understand here. But if you are referring to jetty-servlet-tester, putting the dependency in aaa-parent and putting it as well in the bundles needing it seems to be a duplication to me.

The only benefit I could see is to define versions, and/or common dependencies across bundles. But if a bundle using the parent doesn't need all parents' dependencies, it seems to be adding some extra garbage dependencies for this bundle.

What I believe, and what we discussed a bit last MD-SAL meeting is having all versioning handled by odlparent if the dependency is somewhat shared across project. If the dependency is project specific, then it should be define in the bundle needing it. Finally, if the dependency is shared across multiple all project bundles, then having a parent would make sense.

At least this is how I see things. This way, version, dependencies management is eased

Patch-set: 1