Update patch set 1
Patch Set 1:

| The only benefit I could see is to define versions, and/or common dependencies across bundles.
yes , both of sts and federation are using the same dependency for their unit test, putting this dependency in aaa-parent was much easier to figure out the problem and update the dependency.
the question is, if we delete the aaa-parent and then pushed common dependency across bundles to odlparent does that will screw other projects ? also if we have a station as the last dependency conflict, we will have to  check the odlparent dependencies which is much harder to figure it out.

Patch-set: 1