Enhance idmtool to allow disabling https certificate verification
Adds the capability to disable https certificate verification through
the "-k" or "--insecure" flag.  This vernacular was chosen to closely
mimic curl's interface.  If this mode is enabled, then an appropriate
warning message is printed to make it painfully obvious that HTTPS
certificates are not verified.  This behavior is completely optional,
and is not enabled by default.

Additionally, exception reporting was improved to isolate SSLError(s);
if an SSLError is encountered then it is reported as a possible SSL
issue instead of with the standard "Are you sure the controller is up?"
message.

Change-Id: Ibc138d073d170d76164e928eb0d0cc99f704514c
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>