Bug 6148 AAA project depends on openflowplugin openflowjava
Get rid of build cycle between AAA and openflowplugin.  This is a
temporary fix which just removes the aaa-cert feature/bundle, and
the proper follow up is included in this patch:

https://git.opendaylight.org/gerrit/#/c/36514/

This will get rid of the build dependency on openflow*, and subsequently
ensure the build loop is broken.  The follow up fix rearchitects the AAA
API(s) (newly added in Boron) so that the dependency is not needed.

Change-Id: I040859d8ca29428f146e1b1937d84c288440f92f
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>