Remove stale code from aaa-idmlight bundle
There is a bunch of bash scripts and json included in the aaa-idmlight bundle
that are there for historic reasons only.  These scripts do not reflect the
new data models that have been used for AAA since Beryllium, and thus are confusing
at best.  This change removes this dated code to avoid confusion and clean
up the code base.

Change-Id: Ib698c9823227d9648b65881993276c9c187e3443
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>